chore: add collateral assets to morpho vaults
